refactor: rename google_search.py to web_search.py, optimize search pacing
- Rename to match actual provider (Searlo, not Google)
- Remove redundant logger.warning before raising QuotaExhaustedError
- Move sleep to end of loop so classify/crawl time overlaps pacing
- Make search delay configurable via DISCOVERY_SEARCH_DELAY env var
